fix endpoints
