+++ src/interfaces/item-category-age.ts
@@ -1,0 +1,11 @@
+import { ItemCategory } from "./item-category";
+
+export class ItemCategoryAge implements ItemCategory {
+    itemId: string;
+    categoryAge: string;
+  
+    constructor(itemId: string, categoryAge: string) {
+      this.itemId = itemId;
+      this.categoryAge = categoryAge;
+    }
+}
