+++ src/interfaces/item-category.ts
@@ -3,24 +3,4 @@ export interface ItemCategory {
     categoryAge?: string;
     categoryDiscipline?: string;
 }
-
-export class ItemCategoryDiscipline implements ItemCategory {
-    itemId: string;
-    categoryDiscipline: string;
-  
-    constructor(itemId: string, categoryDiscipline: string) {
-      this.itemId = itemId;
-      this.categoryDiscipline = categoryDiscipline;
-    }
-}
-
-export class ItemCategoryAge implements ItemCategory {
-    itemId: string;
-    categoryAge: string;
-  
-    constructor(itemId: string, categoryAge: string) {
-      this.itemId = itemId;
-      this.categoryAge = categoryAge;
-    }
-}
